Refactor the change for two separate fd_filestat_get_impl
