+++ src/hostcalls_impl/fs.rs
@@ -4,7 +4,7 @@ use crate::fdentry::{Descriptor, FdEntry};
 use crate::memory::*;
 use crate::sys::{errno_from_host, host_impl, hostcalls_impl};
 use crate::{host, wasm32, Result};
-use log::{debug, trace};
+use log::trace;
 use std::io::{self, Read, Seek, SeekFrom, Write};
 
 pub(crate) fn fd_close(wasi_ctx: &mut WasiCtx, fd: wasm32::__wasi_fd_t) -> Result<()> {
@@ -730,53 +730,11 @@ pub(crate) fn fd_filestat_get(
         .get_fd_entry(fd, 0, 0)
         .and_then(|fe| fe.fd_object.descriptor.as_file())?;
 
-    let host_filestat = fd_filestat_get_impl(fd)?;
+    let host_filestat = hostcalls_impl::fd_filestat_get_impl(fd)?;
 
     trace!("     | *filestat_ptr={:?}", host_filestat);
 
     enc_filestat_byref(memory, filestat_ptr, host_filestat)
-}
-
-fn fd_filestat_get_impl(file: &std::fs::File) -> Result<host::__wasi_filestat_t> {
-    use std::convert::TryInto;
-    use std::time::{SystemTime, UNIX_EPOCH};
-
-    fn convert_err(e: io::Error) -> host::__wasi_errno_t {
-        debug!("fd_filestat_get: os error: {}", e);
-        e.raw_os_error().map_or(host::__WASI_EIO, errno_from_host)
-    }
-
-    fn timestamp(st: SystemTime) -> Result<u64> {
-        st.duration_since(UNIX_EPOCH)
-            .map_err(|_| host::__WASI_EINVAL)? // date earlier than UNIX_EPOCH
-            .as_nanos()
-            .try_into()
-            .map_err(|_| host::__WASI_EOVERFLOW) // u128 doesn't fit into u64
-    }
-
-    let metadata = file.metadata().map_err(convert_err)?;
-    Ok(host::__wasi_filestat_t {
-        st_dev: hostcalls_impl::device_id(file, &metadata).map_err(convert_err)?,
-        st_ino: hostcalls_impl::file_serial_no(file, &metadata).map_err(convert_err)?,
-        st_nlink: hostcalls_impl::num_hardlinks(file, &metadata)
-            .map_err(convert_err)?
-            .try_into()
-            .map_err(|_| host::__WASI_EOVERFLOW)?, // u64 doesn't fit into u32
-        st_size: metadata.len(),
-        st_atim: metadata
-            .accessed()
-            .map_err(convert_err)
-            .and_then(timestamp)?,
-        st_ctim: hostcalls_impl::change_time(file, &metadata)
-            .map_err(convert_err)?
-            .try_into()
-            .map_err(|_| host::__WASI_EOVERFLOW)?, // i64 doesn't fit into u64
-        st_mtim: metadata
-            .modified()
-            .map_err(convert_err)
-            .and_then(timestamp)?,
-        st_filetype: hostcalls_impl::filetype(&metadata).map_err(convert_err)?,
-    })
 }
 
 pub(crate) fn fd_filestat_set_times(
