+++ src/sys/windows/hostcalls_impl/fs.rs
@@ -3,10 +3,12 @@
 use super::fs_helpers::*;
 use crate::ctx::WasiCtx;
 use crate::fdentry::FdEntry;
+use crate::helpers::systemtime_to_timestamp;
 use crate::sys::errno_from_host;
 use crate::sys::fdentry_impl::determine_type_rights;
 use crate::sys::host_impl;
 use crate::{host, Result};
+use std::convert::TryInto;
 use std::fs::{File, Metadata};
 use std::io::{self, Seek, SeekFrom};
 use std::os::windows::fs::FileExt;
@@ -193,7 +195,38 @@ pub(crate) fn change_time(file: &File, _metadata: &Metadata) -> io::Result<i64> 
     winx::file::change_time(file)
 }
 
-pub(crate) fn filetype(metadata: &Metadata) -> io::Result<host::__wasi_filetype_t> {
+pub(crate) fn fd_filestat_get_impl(file: &std::fs::File) -> Result<host::__wasi_filestat_t> {
+    fn convert_err(e: io::Error) -> host::__wasi_errno_t {
+        log::debug!("fd_filestat_get: os error: {}", e);
+        e.raw_os_error().map_or(host::__WASI_EIO, errno_from_host)
+    }
+
+    let metadata = file.metadata().map_err(convert_err)?;
+    Ok(host::__wasi_filestat_t {
+        st_dev: device_id(file, &metadata).map_err(convert_err)?,
+        st_ino: file_serial_no(file, &metadata).map_err(convert_err)?,
+        st_nlink: num_hardlinks(file, &metadata)
+            .map_err(convert_err)?
+            .try_into()
+            .map_err(|_| host::__WASI_EOVERFLOW)?, // u64 doesn't fit into u32
+        st_size: metadata.len(),
+        st_atim: metadata
+            .accessed()
+            .map_err(convert_err)
+            .and_then(systemtime_to_timestamp)?,
+        st_ctim: change_time(file, &metadata)
+            .map_err(convert_err)?
+            .try_into()
+            .map_err(|_| host::__WASI_EOVERFLOW)?, // i64 doesn't fit into u64
+        st_mtim: metadata
+            .modified()
+            .map_err(convert_err)
+            .and_then(systemtime_to_timestamp)?,
+        st_filetype: filetype(&metadata).map_err(convert_err)?,
+    })
+}
+
+fn filetype(metadata: &Metadata) -> io::Result<host::__wasi_filetype_t> {
     let ftype = metadata.file_type();
     let ret = if ftype.is_file() {
         host::__WASI_FILETYPE_REGULAR_FILE
